update metamap client with same change and fix format

+++ docker/as/client/required/scripts/metamap_client.groovy
@@ -30,8 +30,8 @@ def group = new DefaultPGroup(8);
 URI sourceUri;
 def scriptDir = new File(sourceUri).parent;
 
-def batchBegin = env["BATCH_BEGIN"] ?: 0;
-def batchEnd = env["BATCH_END"] ?: 9999;
+def batchBegin = (env["BATCH_BEGIN"] ?: 0) as Integer;
+def batchEnd = (env["BATCH_END"] ?: 9999) as Integer;
 def batchOffset = env["BATCH_OFFSET"] ?:  new Random().nextInt(batchEnd - batchBegin);
 
 def dataSource = new BasicDataSource();
